+++ src/renderer/vulkan/vulkan_renderer/mod.rs
@@ -917,7 +917,7 @@ impl Renderer for VulkanRenderer {
   fn load_image_with_staging_initialization(
     &mut self, pixels: impl ImageData, magnification_filter: MagnificationMinificationFilter,
     minification_filter: MagnificationMinificationFilter, address_x: TextureAddressMode,
-    address_y: TextureAddressMode, address_z: TextureAddressMode, mip_levels: u8,
+    address_y: TextureAddressMode, address_z: TextureAddressMode, mip_levels: u32,
   ) -> SarektResult<BufferImageHandle<VulkanBufferFunctions>> {
     Ok(
       BufferImageStore::load_image_with_staging_initialization(
